CNEO: implement CNEO energy and force
Implement energy and force for constrained nuclear-electronic orbital
method. `POTENTIAL CNEO` triggers CNEO calculations for certain atom
kinds (hydrogen is the most common, but no limit in principle) and
nuclear basis functions are supplied by `BASIS_SET NUC basis_name`. PB
series basis sets for proton are provided in `data/NUCLEAR_BASIS_SETS`.
Note that GAPW is mandatory for CNEO calculations to be successful, and
pseudopotentials can be used for classical nuclei.

For electrons k-points can be enabled, but because quantum nuclei are
very localized, for the quantum nuclear equation, a Gamma-point
approximation is always used, regardless of the electronic k-point
setting.